add float window for quick clean memory and launch app

+++ CommonDevLib/src/com/rarnu/devlib/component/intf/FloatWindowListener.java
@@ -3,7 +3,7 @@ package com.rarnu.devlib.component.intf;
 import android.view.View;
 
 public interface FloatWindowListener {
-    void onPositionChanged(View v, float x, float y);
+    void onPositionChanged(View v, int x, int y);
     void onFloatWindowClick();
     void onFloatWindowLongClick();
 }
